Port Phase 5: runtime mixin-apply + init fixes (dev client reaches title)
Runtime-only failures (invisible at build time) surfaced via ./gradlew runClient:
- LivingEntityMixin: applyDamage gained ServerWorld param; +getBaseDimensions
  inject moved here from PlayerEntityMixin (declared on LivingEntity, not Player)
- WorldChunkMixin: setBlockState 3rd param boolean -> int (descriptor)
- WindowMixin: scaleFactor double -> int; CameraMixin update(BlockView->World);
  GameRendererMixin getFov ->float + hud-render retarget; RenderTickCounter(+Accessor)
  field renames (tickProgress/dynamicDeltaTicks); KeyboardMixin KeyInput record
- BBSMod: game rule id "bbsEditing" -> "bbs_editing" (Identifier path must be [a-z0-9._-])
- BBSModClient: register KeyBinding.Category once (was re-registered per key -> crash)
- KeyboardInputMixin: drop @Shadow on inherited Input.movementVector (access via extends)

Dev client now loads on 1.21.11 and reaches the title screen.
Remaining: world-join / in-world render paths (next).

+++ src/client/java/mchorse/bbs_mod/mixin/client/WindowMixin.java
@@ -31,7 +31,7 @@ public class WindowMixin
     private int scaledHeight;
 
     @Shadow
-    private double scaleFactor;
+    private int scaleFactor;
 
     @Inject(method = "getWidth", at = @At("HEAD"), cancellable = true)
     public void onGetWidth(CallbackInfoReturnable<Integer> info)
@@ -74,7 +74,7 @@ public class WindowMixin
     {
         if (BBSRendering.canReplaceFramebuffer())
         {
-            info.setReturnValue((int) (BBSRendering.getVideoWidth() / this.scaleFactor * BBSModClient.getOriginalFramebufferScale()));
+            info.setReturnValue((int) (BBSRendering.getVideoWidth() / (double) this.scaleFactor * BBSModClient.getOriginalFramebufferScale()));
         }
     }
 
@@ -83,7 +83,7 @@ public class WindowMixin
     {
         if (BBSRendering.canReplaceFramebuffer())
         {
-            info.setReturnValue((int) (BBSRendering.getVideoHeight() / this.scaleFactor * BBSModClient.getOriginalFramebufferScale()));
+            info.setReturnValue((int) (BBSRendering.getVideoHeight() / (double) this.scaleFactor * BBSModClient.getOriginalFramebufferScale()));
         }
     }
 }
